+++ src/components/layout/AdminLayout.tsx
@@ -16,7 +16,9 @@ import {
   Church,
   Clock,
   AudioLines,
-  Rss
+  Rss,
+  AlarmClockCheck,
+  Repeat2
 } from 'lucide-react';
 import { useAuth } from '../../contexts/AuthContext';
 import { Button } from '../common/Button';
@@ -49,7 +51,7 @@ export const AdminLayout: React.FC = () => {
         { name: 'Quotes', href: '/admin/quotes', icon: Quote },   
         { name: 'Posts', href: '/admin/posts', icon: AudioLines },
         { name: 'Bulletins', href: '/admin/bulletins', icon: BookOpen },
-        { name: 'Schedules', href: '/admin/schedules', icon: Clock },
+        { name: 'Schedules', href: '/admin/schedules', icon: AlarmClockCheck},
         { name: 'Members', href: '/admin/members', icon: User },
         { name: 'Users', href: '/admin/users', icon: UserCog },
         { name: 'Giving', href: '/admin/giving', icon: HandHeart },
@@ -139,8 +141,8 @@ export const AdminLayout: React.FC = () => {
                     to="/user"
                     className="flex items-center px-3 py-2 text-sm font-medium text-gray-700 hover:bg-gray-100 rounded-md transition-colors duration-200"
                   >
-                    <User className="mr-3 h-4 w-4" />
-                    User Dashboard
+                    <Repeat2 className="mr-3 h-4 w-4" />
+                    User's View
                   </Link>
                   <Button
                     variant="ghost"
@@ -201,8 +203,8 @@ export const AdminLayout: React.FC = () => {
                 to="/user"
                 className="flex items-center px-3 py-2 text-sm font-medium text-gray-700 hover:bg-gray-100 rounded-md transition-colors duration-200"
               >
-                <User className="mr-3 h-4 w-4" />
-                User Dashboard
+                <Repeat2 className="mr-3 h-4 w-4" />
+                User's View
               </Link>
               <Button
                 variant="ghost"
